Add init phase timer on DKGProcess
The init phase timer will only run once for fetching phase node from sharders when the DKGProcess is started. Without the timer, we will wait for 5 seconds, then start to fetch the phase node. This could fix the problem that the first round of view change always failed and need to restart due to some state are not initialized.