Update getMinerNode to get miner node by id from state
Previously, if it could not find the node by id from state, it will then check the minersList and return node from it. This would confuse people, as if the node only exists in minersList, when people try to update the node, the 'node not found' error will be returned, which is weird.